Move Controllers back out of API
They're not part of the API, though they could be and we might address
that complexity if / when it occurs, but trying to do it now while there
aren't any particular plans for it is unnecssary. ADL is unlikely to be
complex enough that it would be hard to refactor everything into being
able to support an external API
