+++ app/Http/Controllers/CreatureTypesController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Models\Data\CreatureType;
 use Illuminate\Http\Request;
